Move MITAardvark format validation method into validator module
